requirements.txt
supabase
flask
python-dotenv
flask_sqlalchemy
gunicorn
Markdown
# compatible with python version 3.12.5 and 3.11.7
